+++ src/format/src/shared/index.ts
@@ -1,0 +1,1 @@
+export { countTrailingBlankLines, getNextNonWhitespaceCharacter } from "./layout-helpers.js";
