+++ src/format/test/format-component-dependencies.test.ts
@@ -15,7 +15,15 @@ void test("dependency bundle is frozen and exposes expected contract keys", () =
 
     assert.deepStrictEqual(
         Object.keys(dependencyBundle).toSorted(),
-        ["LogicalOperatorsStyle", "gmlParserAdapter", "handleComments", "print", "printComment"].toSorted()
+        [
+            "LogicalOperatorsStyle",
+            "countTrailingBlankLines",
+            "getNextNonWhitespaceCharacter",
+            "gmlParserAdapter",
+            "handleComments",
+            "print",
+            "printComment"
+        ].toSorted()
     );
 });
